feat(core): implement chunk based data retrieval strategy
This strategy does a HEAD request to the requested path, parses out the root-transaction-id and corresponding offset. It then constructs a ReadableStream that fetches chunks using offsets from the /chunk/<offset>/data endpoint and streams that into a constructed Response object that is returned to the caller. This simulates a contiguous data request, but done so via chunks!